Updated tests to pass again.
Found some device dependent problems in the tests of the theme
rendering (when monitor resolution changes, so does font size (in
pixels)). This meant that the tests failed (as they are comparisons
with golden-bitmaps). Have updated the tests to do a more "approximate
comparison", with a low tolerance.  This is slow currently, could do
with optimising.

bzr-revno: 411