- It works! - Added text prediction for book names - New script for reference_combined - Also on previous: Added mininium search lengt of 3 characters for text to prevent short keywords from crashing OLP. - Added new icon to code but still using reference icon since it requires packing in order to work.
To DO:
a) New icon, pack it | How?
b) Added stupid error message if nothing is found, make it less stupid.
c) Add example verses to "Reference not found" error. ( and Use this as base for b) 

d) Possibly work on text prediction for Text based search but probably leave it for an another branch.